+++ tests/trainer/connectors/test_callback_connector.py
@@ -66,8 +66,8 @@ class StatefulCallback1(Callback):
         self._other = other
 
     @property
-    def state_id(self):
-        return self._generate_state_id(unique=self._unique)
+    def state_key(self):
+        return self._generate_state_key(unique=self._unique)
 
     def on_save_checkpoint(self, *args):
         return {"content1": self._unique}